requirements.txt
langchain==0.0.188
numpy==1.24.3
openai==1.13.3
playsound==1.3.0
Requests==2.31.0
scipy==1.10.1
sounddevice==0.4.6
openai-whisper==20231117
python-dotenv==1.0.1
pydub==0.25.1